refactor: remove fragment
