+++ simaple/simulate/component/complex_skill.py
@@ -2,8 +2,8 @@ from typing import Optional
 
 from simaple.core.base import Stat
 from simaple.simulate.component.base import ReducerState, reducer_method, view_method
+from simaple.simulate.component.entity import CooldownState, DurationState
 from simaple.simulate.component.skill import SkillComponent
-from simaple.simulate.component.state_fragment import CooldownState, DurationState
 from simaple.simulate.component.trait.impl import (
     DurableTrait,
     InvalidatableCooldownTrait,
