+++ simaple/simulate/component/specific/archmagefb.py
@@ -1,11 +1,7 @@
 from simaple.simulate.base import Entity
 from simaple.simulate.component.base import ReducerState, reducer_method, view_method
+from simaple.simulate.component.entity import CooldownState, IntervalState, StackState
 from simaple.simulate.component.skill import SkillComponent
-from simaple.simulate.component.state_fragment import (
-    CooldownState,
-    IntervalState,
-    StackState,
-)
 from simaple.simulate.component.trait.impl import (
     CooldownValidityTrait,
     TickEmittingTrait,
