+++ simaple/simulate/component/specific/archmagetc.py
@@ -5,12 +5,8 @@ from simaple.simulate.component.base import (
     reducer_method,
     view_method,
 )
+from simaple.simulate.component.entity import CooldownState, IntervalState, StackState
 from simaple.simulate.component.skill import SkillComponent
-from simaple.simulate.component.state_fragment import (
-    CooldownState,
-    IntervalState,
-    StackState,
-)
 from simaple.simulate.component.trait.impl import (
     CooldownValidityTrait,
     StartIntervalWithoutDamageTrait,
