+++ simaple/simulate/component/specific/magician.py
@@ -2,8 +2,8 @@ from typing import Optional
 
 from simaple.core.base import Stat
 from simaple.simulate.component.base import ReducerState, reducer_method, view_method
+from simaple.simulate.component.entity import CooldownState, DurationState
 from simaple.simulate.component.skill import SkillComponent
-from simaple.simulate.component.state_fragment import CooldownState, DurationState
 from simaple.simulate.component.trait.impl import CooldownValidityTrait, DurableTrait
 from simaple.simulate.component.view import Running
 from simaple.simulate.global_property import Dynamics
