+++ tests/simulate/test_save_load.py
@@ -1,5 +1,5 @@
 from simaple.simulate.base import ConcreteStore
-from simaple.simulate.component.state_fragment import (
+from simaple.simulate.component.entity import (
     CooldownState,
     DurationState,
     IntervalState,
